requirements.txt
numpy
torch==1.7.1+cpu
streamlit==0.77.0
Pillow==8.1.1
